renamed test file, added correct license headers
