+++ nav2_costmap_2d/include/nav2_costmap_2d/asymmetric_inflation_layer.hpp
@@ -1,4 +1,4 @@
-// Copyright 2026 Duatic AG
+// Copyright (c) 2026
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
